backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
SpeechRecognition
pydub
python-multipart
google-generativeai
python-dotenv